+++ Sources/PowerSync/Kotlin/KotlinPowerSyncDatabaseImpl.swift
@@ -52,12 +52,21 @@ final class KotlinPowerSyncDatabaseImpl: PowerSyncDatabaseProtocol {
         )
 
         let resolvedOptions = options ?? ConnectOptions()
+        let useWebsockets = switch (resolvedOptions.connectionMethod) {
+            case .http: false
+            case .webSocket: true
+        }
 
         try await kotlinDatabase.connect(
             connector: connectorAdapter,
             crudThrottleMs: Int64(resolvedOptions.crudThrottle * 1000),
             retryDelayMs: Int64(resolvedOptions.retryDelay * 1000),
-            params: resolvedOptions.params.mapValues { $0.toKotlinMap() }
+            params: resolvedOptions.params.mapValues { $0.toKotlinMap() },
+            options: createSyncOptions(
+                newClient: resolvedOptions.newClientImplementation,
+                webSocket: useWebsockets,
+                userAgent: "PowerSync Swift SDK"
+            )
         )
     }