+++ Sources/PowerSync/Protocol/PowerSyncDatabaseProtocol.swift
@@ -34,7 +34,22 @@ public struct ConnectOptions {
     /// ]
     /// ```
     public var params: JsonParam
+    
+    /// Uses a new sync client implemented in Rust instead of the one implemented in Kotlin.
+    ///
+    /// The new client is more efficient and will become the default in the future, but is still marked as experimental for now.
+    /// We encourage interested users to try the new client.
+    @_spi(PowerSyncExperimental)
+    public var newClientImplementation: Bool
 
+    /// The connection method used to connect to the Powersync service.
+    ///
+    /// The default method is ``ConnectionMethod/http``. Using ``ConnectionMethod/webSocket(_:)`` can
+    /// improve performance as a more efficient binary protocol is used. However, using the websocket connection method
+    /// requires enabling ``ConnectOptions/newClientImplementation``.
+    @_spi(PowerSyncExperimental)
+    public var connectionMethod: ConnectionMethod
+    
     /// Initializes a `ConnectOptions` instance with optional values.
     ///
     /// - Parameters:
@@ -49,9 +64,32 @@ public struct ConnectOptions {
         self.crudThrottle = crudThrottle
         self.retryDelay = retryDelay
         self.params = params
+        self.newClientImplementation = false
+        self.connectionMethod = .http
+    }
+    
+    /// Initializes a ``ConnectOptions`` instance with optional values, including experimental options.
+    @_spi(PowerSyncExperimental)
+    public init(
+        crudThrottle: TimeInterval = 1,
+        retryDelay: TimeInterval = 5,
+        params: JsonParam = [:],
+        newClientImplementation: Bool = false,
+        connectionMethod: ConnectionMethod = .http
+    ) {
+        self.crudThrottle = crudThrottle
+        self.retryDelay = retryDelay
+        self.params = params
+        self.newClientImplementation = newClientImplementation
+        self.connectionMethod = connectionMethod
     }
 }
 
+@_spi(PowerSyncExperimental)
+public enum ConnectionMethod {
+    case http
+    case webSocket
+}
 
 /// A PowerSync managed database.
 ///